Add create http headers helper
