+++ tests/util_test.py
@@ -3,6 +3,7 @@ import sys
 import mock
 
 from py_zipkin import util
+from py_zipkin.util import ZipkinAttrs
 
 
 @mock.patch("py_zipkin.util.random.getrandbits", autospec=True)
@@ -44,3 +45,39 @@ def test_signed_int_to_unsigned_hex():
                 util.signed_int_to_unsigned_hex(-5270423489115668655)
                 == "b6dbb1c2b362bf51"
             )
+
+
+@mock.patch("py_zipkin.util.generate_random_128bit_string", autospec=True)
+@mock.patch("py_zipkin.util.generate_random_64bit_string", autospec=True)
+def test_create_attrs_for_span(random_64bit_mock, random_128bit_mock):
+    random_64bit_mock.return_value = "0000000000000042"
+    expected_attrs = ZipkinAttrs(
+        trace_id="0000000000000042",
+        span_id="0000000000000042",
+        parent_span_id=None,
+        flags="0",
+        is_sampled=True,
+    )
+    assert expected_attrs == util.create_attrs_for_span()
+
+    # Test overrides
+    expected_attrs = ZipkinAttrs(
+        trace_id="0000000000000045",
+        span_id="0000000000000046",
+        parent_span_id=None,
+        flags="0",
+        is_sampled=False,
+    )
+    assert expected_attrs == util.create_attrs_for_span(
+        sample_rate=0.0, trace_id="0000000000000045", span_id="0000000000000046",
+    )
+
+    random_128bit_mock.return_value = "00000000000000420000000000000042"
+    expected_attrs = ZipkinAttrs(
+        trace_id="00000000000000420000000000000042",
+        span_id="0000000000000042",
+        parent_span_id=None,
+        flags="0",
+        is_sampled=True,
+    )
+    assert expected_attrs == util.create_attrs_for_span(use_128bit_trace_id=True)
